fix: two lint gates the flatpak commits tripped
lint-qt-includes is the Phase 1f gate: src/qt reaches core state only through
src/interfaces, and util/system.h is on its forbidden list. Putting
IsVolatileFilesystem there meant intro.cpp had to include a core header to ask a
question about a filesystem. Move it to fsbridge, which is where a filesystem
predicate belongs anyway and which src/qt already consumes.

lint-locale-dependence rejects fprintf, via vfprintf. The pre-logging notice now
builds its string with strprintf and writes it with fputs, neither of which is
on that list.

Behaviour is unchanged: an unset XDG_STATE_HOME still resolves under .var/app,
and a default data directory on tmpfs still refuses to start. Both lints pass.
